can now reject leave and expenses

+++ database/users.sql
@@ -60,7 +60,7 @@ CREATE TABLE IF NOT EXISTS `rights` (
 CREATE TABLE IF NOT EXISTS `leave` (
   `id` int (11) NOT NULL AUTO_INCREMENT,
   `user_id` int (11) NOT NULL,
-  `accepted` tinyint(1) DEFAULT 0,
+  `accepted` int(1) DEFAULT 0,
   `start` datetime DEFAULT CURRENT_TIMESTAMP NOT NULL,
   `end` datetime DEFAULT CURRENT_TIMESTAMP NOT NULL,
   `start_time` int(11) DEFAULT 9,
@@ -76,7 +76,7 @@ CREATE TABLE IF NOT EXISTS `leave` (
 CREATE TABLE IF NOT EXISTS `expenses` (
   `id` int (11) NOT NULL AUTO_INCREMENT,
   `user_id` int (11) NOT NULL,
-  `accepted` tinyint(1) DEFAULT 0,
+  `accepted` int(1) DEFAULT 0,
   `year` int(11),
   `month` int(11),
   `amount` decimal(10, 2),
